Scraping/requirements.txt
#Package-----------Version
async-generator==1.10
attrs==21.4.0
Augmentor==0.2.10
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
fonttools==4.33.3
future==0.18.2
h11==0.13.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
opencv-python==4.5.5.64
outcome==1.1.0
packaging==21.3
Pillow==9.1.1
pip==21.2.4
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.5
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wheel==0.37.1
wincertstore==0.2
wsproto==1.1.0